tratamento de erro

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -58,24 +59,35 @@ public class TopicosController {
   }
 
   @GetMapping("/{id}")
-  public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-    Topico topico = topicoRepository.getOne(id);
-    return new DetalhesDoTopicoDto(topico);
+  public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+    Optional<Topico> topico = topicoRepository.findById(id);
+    if (topico.isPresent()) {
+      return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+    }
+    return ResponseEntity.notFound().build();
   }
 
   @PutMapping("/{id}")
   @Transactional
   public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-    Topico topico = form.atualizar(id, topicoRepository);
+    Optional<Topico> topico = topicoRepository.findById(id);
+    if (topico.isPresent()) {
+      Topico topicoAtualizado = form.atualizar(id, topicoRepository);
+      return ResponseEntity.ok(new TopicoDto(topicoAtualizado));
+    }
 
-    return ResponseEntity.ok(new TopicoDto(topico));
+    return ResponseEntity.notFound().build();
   }
 
   @DeleteMapping("/{id}")
   @Transactional
   public ResponseEntity<TopicoDto> remover(@PathVariable Long id) {
-    topicoRepository.deleteById(id);
-    return ResponseEntity.ok().build();
+    Optional<Topico> topico = topicoRepository.findById(id);
+    if (topico.isPresent()) {
+      topicoRepository.deleteById(id);
+      return ResponseEntity.ok().build();
+    }
+    return ResponseEntity.notFound().build();
   }
 
 }
